Adding rough play and stop buttons (#94)
Editor:
* Added rough play and stop buttons.
* Can add and remove script components to any node now.
* Added options 'create directory', 'rename', and 'delete' to asset browser.

Engine:
* Added flag '-ia' to override the 'internal assets' path.
* Build: Create a bin directory to be used by the editor
* Fixed typo of 'game_title' in configure_properties python api function.

+++ editor/src/core/components/component.h
@@ -121,7 +121,7 @@ struct AnimatedSpriteComp : public EditorComponent {
         for (size_t animationIndex = 0; animationIndex < animatedSpriteComponentData->animationCount; animationIndex++) {
             const AnimationData& animData = animatedSpriteComponentData->animations[animatedSpriteComponentData->animationCount];
             EditorAnimation animation = { animData.name, animData.speed, animData.doesLoop };
-            for (size_t frameIndex = 0; frameIndex < animData.frameCount; frameIndex++) {
+            for (size_t frameIndex = 0; (int) frameIndex < animData.frameCount; frameIndex++) {
                 const AnimationFrameData& frameData = animData.animationFrames[frameIndex];
                 const EditorAnimationFrame animationFrame = { frameData.texturePath, frameData.drawSource, frameData.frame };
                 animation.animationFrames.emplace_back(animationFrame);
